backend/requirements.txt
# SmartAM_System 后端依赖
# Python >= 3.8

# ========== Web 框架 ==========
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# ========== 机器学习/深度学习 ==========
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# ========== 图像处理 ==========
opencv-python>=4.8.0
pillow>=10.0.0

# ========== 数据处理和序列化 ==========
pydantic>=2.5.0
pydantic-settings>=2.1.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# ========== 串口通信 ==========
pyserial>=3.5

# ========== HTTP 请求 ==========
requests>=2.31.0

# ========== 工具库 ==========
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0

# ========== 测试 ==========
pytest>=7.4.0
pytest-asyncio>=0.21.0

# ========== 可选：IDS 相机驱动 ==========
# ids-peak 需要单独安装，请参考 IDS 官方文档
# 下载地址：https://en.ids-imaging.com/download.html
# ids-peak>=2.0.0

# ========== 可选：开发工具 ==========
# black>=23.0.0          # 代码格式化
# pylint>=3.0.0          # 代码检查
# mypy>=1.5.0            # 类型检查
